blsct: four small correctness fixes in the proof helpers

+++ src/blsct/arith/elements.cpp
@@ -262,15 +262,15 @@ template MclScalar Elements<MclScalar>::Sum() const;
 template MclG1Point Elements<MclG1Point>::Sum() const;
 
 template <typename T>
-void Elements<T>::ConfirmIndexInsideRange(const uint32_t& index) const
+void Elements<T>::ConfirmIndexInsideRange(const size_t& index) const
 {
     if (index >= m_vec.size()) {
         auto s = strprintf("index %d is out of range [0..%d]", index, m_vec.size() - 1ul);
         throw std::runtime_error(s);
     }
 }
-template void Elements<MclScalar>::ConfirmIndexInsideRange(const uint32_t&) const;
-template void Elements<MclG1Point>::ConfirmIndexInsideRange(const uint32_t&) const;
+template void Elements<MclScalar>::ConfirmIndexInsideRange(const size_t&) const;
+template void Elements<MclG1Point>::ConfirmIndexInsideRange(const size_t&) const;
 
 template <typename T>
 T& Elements<T>::operator[](const size_t& index)
